Ensure share repo is updated when versioning
Currently, it's easy to forget manually pushing to master after
versioning/building. In which case things can be merged remotely, and
when pulling with rebase the version commit ends up being changed and
behind what was actually tagged.

This adds a postversion script which simply pushes to the remote
automatically after the version commit and tag have been added by npm.