Fix timezone issue in expense item form
Dates being added with 00:00 time still carry the timezone set for the
OS/browser, and thus get changed to the previous day when in a positive
offset zone. This change removes the timezone offset entirely when
adding the line item, so that the chosen date is always assumed to be in
UTC time.

fixes #204